b4pull

+++ FintrakBanking.Entities/Models/TBL_LOAN_COLLATERAL_MAPPING.cs
@@ -18,7 +18,7 @@ namespace FintrakBanking.Entities.Models
 
         public int LOANID { get; set; }
 
-        public int LOANSYSTEMTYPEID { get; set; }
+        public short LOANSYSTEMTYPEID { get; set; }
 
         public bool ISRELEASED { get; set; }
